Fix error when using script_runner.py on iOS<14
Running a Python module in Pyto from `Run module (python -m)` results in an error on iOS < 13. This patch adds a `is_watch_script` check before importing `watch`, as it requires the `widgets` module (available in iOS 14+).

A check could be added to `watch.py` before importing `widgets` to check if the current iOS is supported, but I don’t know how to do that.